Track processed chat paths in a separate file
Before, we skipped processed chats if a directory already existed for
it, but we only created directories for chats with images. So we would
read and parse all chats without images no matter what (and actually
also ones with images, but if the condition were moved that would be
fixed). Now we store an array of paths (indexed as a set for matching)
in a plist inside our Caches/Metadata subdirectory.